requirements.txt
selenium
pytest
allure-pytest
python-dotenv
requests